# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Tether-ERP Procurement API
  description: |
    Comprehensive REST API for procurement management system with approval workflows,
    budget constraints, document linking, and multi-stage approval chains.
  version: 1.0.0
  contact:
    name: Tether-ERP Support
    email: support@tether-erp.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development Server
  - url: https://api.tether-erp.com/api/v1
    description: Production Server
...
paths:
...
  /organizations:
    get:
      tags:
        - Organizations
      summary: List user's organizations
      description: Get all organizations the current user belongs to
      operationId: getOrganizations
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Organization"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags:
        - Organizations
      summary: Create organization
      description: Create a new organization
      operationId: createOrganization
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Organization"
        "400":
          $ref: "#/components/responses/BadRequest"
  /organizations/{id}:
    put:
      tags:
        - Organizations
      summary: Update organization
      operationId: updateOrganization
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: Organization updated
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - Organizations
      summary: Delete organization
      operationId: deleteOrganization
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Organization deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{id}/switch:
    post:
      tags:
        - Organizations
      summary: Switch organization
      description: Switch the current active organization context
      operationId: switchOrganization
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Organization switched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        description: New JWT token with updated organization context
        "404":
          $ref: "#/components/responses/NotFound"
  /organization/members:
    get:
      tags:
        - Organization Members
      summary: List organization members
      operationId: getOrganizationMembers
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrganizationMember"

    post:
      tags:
        - Organization Members
      summary: Add organization member
      operationId: addOrganizationMember
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                role:
                  type: string
      responses:
        "201":
          description: Member added
        "400":
          $ref: "#/components/responses/BadRequest"
  /organization/members/{userId}:
    delete:
      tags:
        - Organization Members
      summary: Remove organization member
      operationId: removeOrganizationMember
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member removed
        "404":
          $ref: "#/components/responses/NotFound"
...
  /organization/settings:
    get:
      tags:
        - Organization Members
      summary: Get organization settings
      operationId: getOrganizationSettings
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Organization settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object

    put:
      tags:
        - Organization Members
      summary: Update organization settings
      operationId: updateOrganizationSettings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Settings updated
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, approver, user]
        department:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        logo:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateOrganizationRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 3
        description:
          type: string
        logo:
          type: string
    UpdateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        logo:
          type: string
    OrganizationMember:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        organizationId:
          type: string
          format: uuid
        role:
          type: string
        user:
          $ref: "#/components/schemas/User"
        joinedAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Invalid request body
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Unauthorized
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Resource not found